modularize loss ops
