+++ tfjs-core/src/tests.ts
@@ -38,6 +38,7 @@ import './io/router_registry_test';
 import './io/weights_loader_test';
 import './jasmine_util_test';
 import './kernel_registry_test';
+import './ops/absolute_difference_test';
 import './ops/add_n_test';
 import './ops/add_test';
 import './ops/arithmetic_test';
@@ -55,6 +56,7 @@ import './ops/broadcast_util_test';
 import './ops/clone_test';
 import './ops/compare_ops_test';
 import './ops/complex_ops_test';
+import './ops/compute_weighted_loss_test';
 import './ops/concat_test';
 import './ops/concat_util_test';
 import './ops/confusion_matrix_test';
@@ -66,6 +68,7 @@ import './ops/conv2d_transpose_test';
 import './ops/conv3d_test';
 import './ops/conv3d_transpose_test';
 import './ops/conv_util_test';
+import './ops/cosine_distance_test';
 import './ops/cumsum_test';
 import './ops/depth_to_space_test';
 import './ops/diag_test';
@@ -79,22 +82,25 @@ import './ops/gather_nd_test';
 import './ops/gram_schmidt_test';
 import './ops/greater_equal_test';
 import './ops/greater_test';
+import './ops/hinge_loss_test';
+import './ops/huber_loss_test';
 import './ops/image_ops_test';
 import './ops/in_top_k_test';
 import './ops/leaky_relu_test';
 import './ops/less_equal_test';
 import './ops/less_test';
 import './ops/local_response_normalization_test';
+import './ops/log_loss_test';
 import './ops/logical_and_test';
 import './ops/logical_not_test';
 import './ops/logical_or_test';
 import './ops/logical_xor_test';
-import './ops/loss_ops_test';
 import './ops/lstm_test';
 import './ops/mat_mul_test';
 import './ops/max_pool_3d_test';
 import './ops/max_pool_test';
 import './ops/max_pool_with_argmax_test';
+import './ops/mean_squared_error_test';
 import './ops/moving_average_test';
 import './ops/multinomial_test';
 import './ops/not_equal_test';
@@ -116,9 +122,11 @@ import './ops/reverse_test';
 import './ops/scatter_nd_test';
 import './ops/segment_ops_test';
 import './ops/selu_test';
+import './ops/sigmoid_cross_entropy_test';
 import './ops/signal_ops_test';
 import './ops/slice_test';
 import './ops/slice_util_test';
+import './ops/softmax_cross_entropy_test';
 import './ops/softmax_test';
 import './ops/space_to_batch_nd_test';
 import './ops/sparse_to_dense_test';
